requirement.txt
aiogram
pycryptodome